requirements.txt
camelot-py[cv]
pandas==2.3.0
PySide6==6.9.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
